ADD PROJECTS TO PROJECTS GALLERY
Added calculator to the projects gallery, no javascript yet as need
to review the code beofre adding it to the porfolio.

Added a thumbnail for the etch-a-sketch project and the write up for
it too.

Also added padding-left: 5px to the h3 elements as the text was up against
the left hand side of the element.