SOLVE YOFFSET ISSUE WHEN USING NAVBAR LINKS
Clicking on the links in the nav bar took the user to the correct
location on the page but failed to account for the height of the navbar at the top of the page. This has been solved by adding padding
equal to the height of the nav bar to each of the elements that can be reached through the nav bar.