RESOLVED ISSUES WITH CALCULATOR
Calculator was failing to display the answer when the equals was pressed,
the issue was a failure to pass in a boolean value to the claculate function.

The clear button wasn't working as expected, when clearing the operator
the dispaly answer function wasn't being called and an if statement had
to be added for if num2 is only one character long.